Split off common code
