+++ arrow-fx-coroutines-stream/build.gradle
@@ -13,6 +13,7 @@ dependencies {
     implementation project(":arrow-fx-coroutines")
     implementation project(":arrow-fx-suspend-connection")
 
+    testImplementation project(":arrow-fx-coroutines-test")
     testImplementation "io.kotest:kotest-runner-junit5-jvm:$KOTEST_VERSION" // for kotest framework
     testImplementation "io.kotest:kotest-assertions-core-jvm:$KOTEST_VERSION" // for kotest core jvm assertions
     testImplementation "io.kotest:kotest-property-jvm:$KOTEST_VERSION" // for kotest property test
