+++ arrow-fx-coroutines-stream/src/test/kotlin/arrow/fx/coroutines/stream/ConcurrentlyTest.kt
@@ -71,7 +71,7 @@ class ConcurrentlyTest : StreamSpec(spec = {
     }
 
     "run finalizers of background stream and properly handle exception" {
-      io.kotest.property.checkAll(2000, Arb.stream(Arb.int()), Arb.throwable()) { s, e ->
+      checkAll(Arb.stream(Arb.int()), Arb.throwable()) { s, e ->
         val runnerRun = Atomic(false)
         val finRef = Atomic<List<String>>(emptyList())
         val halt = Promise<Unit>()
