+++ arrow-fx-coroutines-stream/src/test/kotlin/arrow/fx/coroutines/stream/PullTest.kt
@@ -1,5 +1,6 @@
 package arrow.fx.coroutines.stream
 
+import arrow.fx.coroutines.ArrowFxSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.property.Arb
 import io.kotest.property.arbitrary.int
