+++ arrow-fx-coroutines-test/build.gradle
@@ -1,0 +1,19 @@
+plugins {
+    id "org.jetbrains.kotlin.jvm"
+    id "org.jlleitschuh.gradle.ktlint"
+}
+
+apply plugin: 'kotlinx-atomicfu'
+
+apply from: "$SUB_PROJECT"
+apply from: "$DOC_CREATION"
+
+dependencies {
+    api "io.arrow-kt:arrow-core:$VERSION_NAME"
+    api project(":arrow-fx-coroutines")
+    api project(":arrow-fx-coroutines-stream")
+
+    api "io.kotest:kotest-runner-junit5-jvm:$KOTEST_VERSION" // for kotest framework
+    api "io.kotest:kotest-assertions-core-jvm:$KOTEST_VERSION" // for kotest core jvm assertions
+    api "io.kotest:kotest-property-jvm:$KOTEST_VERSION" // for kotest property test
+}
