+++ arrow-fx-coroutines-test/gradle.properties
@@ -1,0 +1,4 @@
+# Maven publishing configuration
+POM_NAME=Arrow-Fx-Coroutines-Test
+POM_ARTIFACT_ID=arrow-fx-coroutines-Test
+POM_PACKAGING=jar
