+++ arrow-fx-coroutines/build.gradle
@@ -12,6 +12,7 @@ dependencies {
     api "io.arrow-kt:arrow-core:$VERSION_NAME"
     implementation project(":arrow-fx-suspend-connection")
 
+    testImplementation project(":arrow-fx-coroutines-test")
     testImplementation "io.kotest:kotest-runner-junit5-jvm:$KOTEST_VERSION" // for kotest framework
     testImplementation "io.kotest:kotest-assertions-core-jvm:$KOTEST_VERSION" // for kotest core jvm assertions
     testImplementation "io.kotest:kotest-property-jvm:$KOTEST_VERSION" // for kotest property test
