+++ arrow-fx-stm/build.gradle
@@ -11,6 +11,7 @@ apply from: "$DOC_CREATION"
 dependencies {
     compile project(":arrow-fx-coroutines")
 
+    testImplementation project(":arrow-fx-coroutines-test")
     testImplementation "io.kotest:kotest-runner-junit5-jvm:$KOTEST_VERSION" // for kotest framework
     testImplementation "io.kotest:kotest-assertions-core-jvm:$KOTEST_VERSION" // for kotest core jvm assertions
     testImplementation "io.kotest:kotest-property-jvm:$KOTEST_VERSION" // for kotest property test
