+++ arrow-fx-stm/src/test/kotlin/arrow/fx/stm/TArrayTest.kt
@@ -1,5 +1,6 @@
 package arrow.fx.stm
 
+import arrow.fx.coroutines.ArrowFxSpec
 import io.kotest.matchers.ints.shouldBeExactly
 import io.kotest.matchers.shouldBe
